ccmtcnvt.c: Handle star-slash inside a C++ comment
It is legal to have */ occur inside a C++-style (//) comment, but
since C-style (/*) comments do not nest, it will prematurely close
the converted comment.  There is no good solution to it.  What I do here
is insert a space between the star and the slash, and print out a note.

Thanks to Ryan (mysterio.ryan@gmail.com) for spotting the issue.

Signed-off-by: Antti-Juhani Kaijanaho <antti-juhani@kaijanaho.fi>